requirements.txt
# discord.py # Removed
# uvloop # Removed (was likely a discord.py dependency)
beautifulsoup4
google-generativeai
requests
# selenium # Assuming not needed by utils? Keep if generate_yt/scrape needs it. Let's remove for now.
cloudscraper
google-genai
# python-dotenv # Removed
